Add separate parameter sets for PR algorithms for compatibility.

+++ src/Application/AppUtils.cpp
@@ -224,8 +224,8 @@ void AppUtils::PrintMatR16(cv::Mat& mat, int value, bool invert, int rowLimit, i
       {
          for (int j = 0; j < cols; j++)
          {
-            float current = mat.at<float>(i, j);
-            printf("%.2lf ", current);
+            uint16_t current = mat.at<uint16_t>(i, j);
+            printf("%hu ", current);
          }
          printf("\n");
       }
@@ -235,7 +235,7 @@ void AppUtils::PrintMatR16(cv::Mat& mat, int value, bool invert, int rowLimit, i
       {
          for (int j = 0; j < cols; j++)
          {
-            float current = mat.at<float>(i, j);
+            uint16_t current = mat.at<uint16_t>(i, j);
             printf("(%d %d): %.2lf\n", i, j, current);
          }
          printf("\n");
